M50 Configuration
This configuration is based on the motorway (M50) scenario from the study "Quantifying the impact of connected and autonomous vehicles on traffic efficiency and safety in mixed traffic" by Maxime Guériau and Ivana Dusparic. The configuration uses SUMO (Simulation of Urban MObility) to model a 7km stretch of the M50 motorway in Ireland, including two major interchanges.This configuration allows for simulating different scenarios with varying penetration rates of CAVs, different traffic demand levels, and enables the evaluation of both traffic efficiency and safety impacts in a motorway setting.